requirements.txt
tensorflow
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
nltk 

